refactor: reorganize entities folder into logical categories
Restructure entities directory for better code organization:

Structure:
- entities/characters/ - character.js, player.js, enemy.js, boss.js, npc.js
- entities/skills/ - skill.js, projectile.js, area-attack.js
- entities/items/ - item.js
- entities/world/ - terrain.js, environment.js
- entities/base-entity.js (root level)

Updated all import paths to reflect new structure:
- Core game imports now reference categorized paths
- Field.js updated to import from entities/world/
- All internal entity cross-references updated
- Maintained proper inheritance chain through relative imports

No functional changes, pure organizational refactor for maintainability.
