+++ public/src/core/game.js
@@ -2,11 +2,11 @@ import * as THREE from 'three';
 import { SceneManager } from './scene-manager.js';
 import { AssetLoader } from './asset-loader.js';
 import { Field } from '../world/field.js';
-import { Player } from '../entities/player.js';
-import { Enemy } from '../entities/enemy.js';
-import { Boss } from '../entities/boss.js';
-import { Npc } from '../entities/npc.js';
-import { Item } from '../entities/item.js';
+import { Player } from '../entities/characters/player.js';
+import { Enemy } from '../entities/characters/enemy.js';
+import { Boss } from '../entities/characters/boss.js';
+import { Npc } from '../entities/characters/npc.js';
+import { Item } from '../entities/items/item.js';
 import { InputController } from '../controls/input-controller.js';
 import { Hud } from '../ui/hud.js';
 import { TitleScreen } from '../ui/title-screen.js';
