+++ public/src/entities/characters/boss.js
