+++ public/src/entities/skills/projectile.js
